refactor(api): extract approvals + TOTP handlers from system.rs (#3749 6/N)
Continues the system.rs sub-domain split begun in #4452 / #4454 / #4455 /
#4458 / #4459. Moves the approval workflow surface — manual create,
get/list, approve/reject/modify, batch resolve, per-session list /
approve_all / reject_all, audit log, pending count — and the TOTP
enrollment surface (setup, confirm, status, revoke) into a sibling
module routes/approvals.rs.

system::router() now delegates these paths via .merge(routes::approvals
::router()). Public route paths and JSON shapes are unchanged.
PaginationParams::effective_{limit,offset} and its consts were bumped to
pub(crate) so the new module can reuse the type via
super::system::PaginationParams. routes/mod.rs gets pub mod approvals
+ pub use approvals::* so utoipa bindings in openapi.rs
(routes::list_approvals, routes::create_approval, routes::get_approval,
routes::approve_request, routes::reject_request) keep resolving without
edits.

system.rs goes from 5939 -> 4451 lines (-1488); new approvals.rs is 1514
lines.

Refs #3749 (6 of N).
